backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
astroid==2.5.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
beanie==1.24.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
databases==0.8.0
distlib==0.3.8
dnspython==2.4.2
ecdsa==0.18.0
EditorConfig==0.12.3
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fake-useragent==1.3.0
Faker==20.1.0
fastapi==0.109.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.0
google==3.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.8.0
jose==1.0.0
jsbeautifier==1.14.11
jwt==1.3.1
kiwisolver==1.4.5
lazy-model==0.2.0
lazy-object-proxy==1.6.0
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mccabe==0.6.1
motor==3.3.2
multidict==6.0.4
mysql-connector-python==9.0.0
numpy==1.26.4
openai==0.28.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
passlib==1.7.4
pendulum==3.0.0
phonenumbers==8.13.26
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyExecJS==1.5.1
PyJWT==2.8.0
pylint==2.7.4
pymongo==4.6.1
PyMySQL==1.1.1
pyparsing==3.1.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyshorteners==1.0.1
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pywifi==1.1.12
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
socialscan==2.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.51
starlette==0.36.3
svgwrite==1.4.3
time-machine==2.14.2
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
Tree==0.2.4
trio==0.23.1
trio-websocket==0.11.1
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.27.0.post1
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.12.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
